requirements-dev.txt
-r requirements.txt
WebOb==1.3.1
WebTest==2.0.14
beautifulsoup4==4.3.2
django-debug-toolbar==1.0.1
django-webtest==1.7.6
factory-boy==2.3.1
flake8==2.1.0
mccabe==0.2.1
pep8==1.4.6
pyflakes==0.7.3
waitress==0.8.8
